refactor: name directory (#1)
