+++ routes/index.js
@@ -1,9 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const controllers = require("../controladores/usuario.js");
+const controllers = require("../handlers/usuario.js");
 
 
-//rutas usuario 
+//routes usuario 
 router.route("/gettodo").get(controllers.getAllTodos)
 router.route("/createTodo").post(controllers.createTodo);
 router.route("/get").get(controllers.getTodo);
