+++ src/public/index.html
@@ -7,13 +7,14 @@
 
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-
-    <link href="style.css" rel="stylesheet">
     
-    <title>MAPA</title>
+    <link href="style.css" rel="stylesheet">
+
+    <title>MUSEO DEL MAR</title>
   </head>
   
-  <body>
+  <body >
+    
     <nav class="navbar navbar-expand-lg navbar fixed-top navbar-dark bg-primary">
         <div class="container-fluid">
           <a class="navbar-brand" href="index.html">Museo del Mar</a>
@@ -29,7 +30,7 @@
                 <a class="nav-link" href="registrarse.html">registrarse</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="iniciar sesion.html">iniciar sesion</a>
+                <a class="nav-link" href="login.html">iniciar sesion</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="mapa.html">mapa</a>
@@ -38,5 +39,21 @@
           </div>
         </div>
       </nav>
+
+      <aside class="sidebar">
+        <div class="descripcion">
+          <label> hola mi nombre es facundo </label>
+        </div>
+      </aside>
+
+      <article class="main">
+      
+      </article>
+
+      <footer class="footer">
+        
+      </footer>
+
   </body>
+
 </html>
