+++ src/public/info.html
@@ -7,14 +7,13 @@
 
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    
-    <link href="style.css" rel="stylesheet">
 
-    <title>MUSEO DEL MAR</title>
+    <link href="style.css" rel="stylesheet">
+    
+    <title>INFORMACION</title>
   </head>
   
-  <body >
-    
+  <body>
     <nav class="navbar navbar-expand-lg navbar fixed-top navbar-dark bg-primary">
         <div class="container-fluid">
           <a class="navbar-brand" href="index.html">Museo del Mar</a>
@@ -30,7 +29,7 @@
                 <a class="nav-link" href="registrarse.html">registrarse</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="iniciar sesion.html">iniciar sesion</a>
+                <a class="nav-link" href="login.html">iniciar sesion</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="mapa.html">mapa</a>
@@ -39,21 +38,5 @@
           </div>
         </div>
       </nav>
-
-      <aside class="sidebar">
-        <div class="descripcion">
-          <label> hola mi nombre es facundo </label>
-        </div>
-      </aside>
-
-      <article class="main">
-      
-      </article>
-
-      <footer class="footer">
-        
-      </footer>
-
   </body>
-
 </html>
