+++ src/public/login.html
@@ -10,37 +10,22 @@
 
     <link href="style.css" rel="stylesheet">
 
-    <title>REGISTRARSE</title>
+    <title>INICIAR SESION</title>
   </head>
   
   <body>
-    <form action="/" class="login">
-
+  
+    <form class="login">
       <div class="mb-3">
         <label for="exampleInputEmail1" class="form-label">Dirección de correo electrónico</label>
-        <input type="email" name="email" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp">
+        <input type="email" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp">
         <div id="emailHelp" class="form-text">We'll never share your email with anyone else.</div>
       </div>
-
-      <div class="mb-3">
-        <label  class="form-label">Nombre</label>
-        <input type="nombre" class="form-control" name="nombre" id="exampleInputNombre">
-      </div>
-
-      <div class="mb-3">
-        <label class="form-label">Apellido</label>
-        <input type="Apellido" class="form-control" name="apellido" id="exampleInputApellido">
-      </div>
-
       <div class="mb-3">
         <label for="exampleInputPassword1" class="form-label">Clave</label>
-        <input type="password" class="form-control"  name="password" id="exampleInputPassword1">
+        <input type="password" class="form-control" id="exampleInputPassword1">
       </div>
-     
-     
-
-
-
+      
       <button type="submit" class="btn btn-primary">Submit</button>
     </form>
 
@@ -59,7 +44,7 @@
                 <a class="nav-link" href="registrarse.html">registrarse</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="iniciar sesion.html">iniciar sesion</a>
+                <a class="nav-link" href="login.html">iniciar sesion</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="mapa.html">mapa</a>
