+++ src/public/mapa.html
@@ -9,26 +9,11 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
 
     <link href="style.css" rel="stylesheet">
-
-    <title>INICIAR SESION</title>
+    
+    <title>MAPA</title>
   </head>
   
   <body>
-  
-    <form class="login">
-      <div class="mb-3">
-        <label for="exampleInputEmail1" class="form-label">Dirección de correo electrónico</label>
-        <input type="email" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp">
-        <div id="emailHelp" class="form-text">We'll never share your email with anyone else.</div>
-      </div>
-      <div class="mb-3">
-        <label for="exampleInputPassword1" class="form-label">Clave</label>
-        <input type="password" class="form-control" id="exampleInputPassword1">
-      </div>
-      
-      <button type="submit" class="btn btn-primary">Submit</button>
-    </form>
-
     <nav class="navbar navbar-expand-lg navbar fixed-top navbar-dark bg-primary">
         <div class="container-fluid">
           <a class="navbar-brand" href="index.html">Museo del Mar</a>
@@ -44,7 +29,7 @@
                 <a class="nav-link" href="registrarse.html">registrarse</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="iniciar sesion.html">iniciar sesion</a>
+                <a class="nav-link" href="login.html">iniciar sesion</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="mapa.html">mapa</a>
